@@ -11,29 +11,62 @@
   "dependencies": {
     "@expo/vector-icons": "^10.0.0",
     "color": "^3.1.2",
-    "expo": "^35.0.0",
+    "expo": "^36.0.0",
+    "expo-font": "~8.0.0",
     "expokit": "^32.0.3",
     "lodash": "^4.17.11",
     "moment": "^2.24.0",
-    "native-base": "^2.12.1",
+    "native-base": "^2.13.8",
     "prop-types": "^15.6.2",
-    "react": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react": "16.9.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.1.tar.gz",
     "react-native-custom-checkbox": "^1.5.2",
+    "react-native-dotenv": "^0.2.0",
+    "react-native-gesture-handler": "^1.5.3",
     "react-native-keyboard-aware-scroll-view": "^0.9.1",
     "react-native-masked-text": "^1.12.2",
     "react-native-modal": "^11.5.3",
     "react-native-picker-select": "^6.3.3",
+    "react-native-screens": "^2.0.0-beta.2",
+    "react-native-svg-transformer": "^0.14.3",
     "react-navigation": "^3.0.9",
     "react-navigation-redux-helpers": "^2.0.9",
+    "react-navigation-stack": "^2.1.1",
+    "react-navigation-tabs": "^2.7.0",
     "react-redux": "^6.0.0",
     "react-style-proptype": "^3.2.2",
+    "reactn": "^2.2.4",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "babel-preset-expo": "^7.0.0",
-    "expo-cli": "^2.16.1"
+    "@babel/core": "^7.6.2",
+    "@babel/runtime": "^7.6.2",
+    "@react-native-community/eslint-config": "^0.0.5",
+    "@types/es6-promisify": "^6.0.0",
+    "@types/jest": "^24.0.23",
+    "@types/react": "~16.8.23",
+    "@types/react-native": "~0.57.65",
+    "@types/react-native-datepicker": "~1.7.0",
+    "@types/react-native-dialog": "~5.6.0",
+    "@types/react-native-dotenv": "^0.2.0",
+    "@types/react-native-vector-icons": "^6.4.4",
+    "@typescript-eslint/eslint-plugin": "^2.8.0",
+    "@typescript-eslint/parser": "^2.7.0",
+    "babel-jest": "^24.9.0",
+    "babel-preset-expo": "^8.0.0",
+    "eslint": "^6.7.2",
+    "eslint-config-prettier": "^6.5.0",
+    "eslint-plugin-jest": "^23.1.1",
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.16.0",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "expo-cli": "^2.21.2",
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "prettier": "^1.19.1",
+    "react-test-renderer": "16.9.0",
+    "typescript": "^3.7.3"
   },
   "private": true
 }
