@@ -1,11 +1,11 @@
 /**
-* Includes API urls, third party keys/secret, etc to be used on app
-*/
+ * Includes API urls, third party keys/secret, etc to be used on app
+ */
 export default {
   // will use depending on the mode
   API_URL: {
-    DEV: 'http://ugg-2957.botics.co/',
-    // DEV: 'http://127.0.0.1:8000/',
-    PROD: '',
-  },
+    DEV: 'https://ugg-2957.botics.co/',
+    // DEV: 'http://localhost:8000/',
+    PROD: ''
+  }
 };
