@@ -3,7 +3,6 @@ import { StackActions, NavigationActions } from 'react-navigation';
 
 import Config from '../constants/config';
 
-
 /**
  * Global Util Functions
  *
@@ -15,9 +14,17 @@ const Utils = {
   // serverUrl: Config.API_URL.PROD,
 
   /**
-    * Checks if the input string is a valid email address
-    */
-  isValidEmail: (emailStr) => {
+   * Checks if the input string is a valid email address
+   */
+  capitaliseFirstChar: str => {
+    return str && str.length === 1 ? str.toUpperCase() : str;
+  },
+  phoneInputRefInvokeMethod: (itemRef, method = 'focus') => {
+    if (itemRef && typeof itemRef[method] === 'function') {
+      itemRef[method]();
+    }
+  },
+  isValidEmail: emailStr => {
     const reg = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
     if (reg.test(emailStr) === false) {
       return false;
@@ -29,15 +36,18 @@ const Utils = {
   numberformatter: (number, decimals, decimalPoint, thousandsSeparator) => {
     const num = !Number.isFinite(+number) ? 0 : +number;
     const precision = !Number.isFinite(+decimals) ? 0 : Math.abs(decimals);
-    const separator = (typeof thousandsSeparator === 'undefined') ? ',' : thousandsSeparator;
-    const decimal = (typeof decimalPoint === 'undefined') ? '.' : decimalPoint;
+    const separator =
+      typeof thousandsSeparator === 'undefined' ? ',' : thousandsSeparator;
+    const decimal = typeof decimalPoint === 'undefined' ? '.' : decimalPoint;
 
     const toFixedFix = (n, prec) => {
       const k = 10 ** prec;
       return Math.round(n * k) / k;
     };
 
-    const s = (precision ? toFixedFix(num, precision) : Math.round(num)).toString().split('.');
+    const s = (precision ? toFixedFix(num, precision) : Math.round(num))
+      .toString()
+      .split('.');
 
     if (s[0].length > 3) {
       s[0] = s[0].replace(/\B(?=(?:\d{3})+(?!\d))/g, separator);
@@ -52,45 +62,39 @@ const Utils = {
   },
 
   // convert true/false string to bool
-  stringToBool: string => (
-    string === 'true'
-  ),
+  stringToBool: string => string === 'true',
 
   // opens a WebView from a URL
-  openURL: (url) => {
-    Linking.canOpenURL(url).then((supported) => {
-      if (supported) {
-        Linking.openURL(url);
-      }
-    }).catch(error => Utils.showAlertError(error));
+  openURL: url => {
+    Linking.canOpenURL(url)
+      .then(supported => {
+        if (supported) {
+          Linking.openURL(url);
+        }
+      })
+      .catch(error => Utils.showAlertError(error));
   },
 
   // resets navigation with the selected route
-  resetNavigationForRoute: route => (
+  resetNavigationForRoute: route =>
     StackActions.reset({
       index: 0,
-      actions: [
-        NavigationActions.navigate({ routeName: route }),
-      ],
-    })
-  ),
+      actions: [NavigationActions.navigate({ routeName: route })]
+    }),
 
   // shows an error alert
-  showAlertError: (message) => {
-    Alert.alert(
-      'Error',
-      message,
-      [
-        { text: 'OK' },
-      ],
-    );
+  showAlertError: message => {
+    Alert.alert('Error', message, [{ text: 'OK' }]);
   },
 
   // creates a list with a range of numbers starting from 0
   range: end => [...Array(end)].map((_, index) => index),
 
   // creates a list with a range of numbers with start and end values
-  rangeFrom: (start, end) => (Array(end - start).fill().map((_, index) => index + start)),
+  rangeFrom: (start, end) =>
+    Array(end - start)
+      .fill()
+      .map((_, index) => index + start)
 };
 
 export default Utils;
